Seeder de tipos de accion,Correcion de plan de accion y policias del mismo

+++ app/Http/Controllers/API/ActionPlanAction/ActionPlanActionController.php
@@ -8,8 +8,9 @@ use App\Http\Requests\ActionPlanAction\StoreActionPlanActionRequest;
 use App\Http\Requests\ActionPlanAction\UpdateActionPlanActionRequest;
 use App\Http\Requests\ActionPlanAction\PartialUpdateActionPlanActionRequest;
 use App\Models\ActionPlanAction\ActionPlanAction;
+use App\Models\ActionPlan\ActionPlan;
 use App\Services\ActionPlanAction\ActionPlanActionService;
-// use App\Policies\AccessActionPlanActionPolicy;
+use App\Policies\AccessActionPlanPolicy;
 
 class ActionPlanActionController extends Controller
 {
@@ -43,12 +44,12 @@ class ActionPlanActionController extends Controller
             return ResponseFormatter::error('Registro no encontrado', 404);
         }
 
-        // if (!(new AccessActionPlanActionPolicy())->access($actionPlanAction)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para ver este registro',
-        //         403
-        //     );
-        // }
+        if (!(new AccessActionPlanPolicy())->access($actionPlanAction->actionPlan)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para ver este registro',
+                403
+            );
+        }
 
         $response = $this->service->getById($id);
 
@@ -63,8 +64,42 @@ class ActionPlanActionController extends Controller
         );
     }
 
+    public function getActionForActionPlan(string $id)
+    {
+        $actionPlan = ActionPlan::find($id);
+
+        if (!$actionPlan) {
+            return ResponseFormatter::error('Registro no encontrado', 404);
+        }
+
+        if (!(new AccessActionPlanPolicy())->access($actionPlan)) {
+            return ResponseFormatter::error('Usted no tiene autorización para ver estos registros',403);
+        }
+
+        $response = $this->service->getActionForActionPlan($id);
+
+        if ($response['error']) {
+            return ResponseFormatter::error($response['message'], $response['code']);
+        }
+
+        return ResponseFormatter::success($response['message'], $response['code'], $response['data'] ?? [], $response['paginate']);
+    }
+
     public function store(StoreActionPlanActionRequest $request)
     {
+        $actionPlan = ActionPlan::find($request->action_plan_id);
+
+        if (!$actionPlan) {
+            return ResponseFormatter::error('Plan de accion no encontrado', 404);
+        }
+
+        if (!(new AccessActionPlanPolicy())->access($actionPlan)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para crear este registro',
+                403
+            );
+        }
+
         $data = $request->validated();
 
         $response = $this->service->create($data);
@@ -88,12 +123,12 @@ class ActionPlanActionController extends Controller
             return ResponseFormatter::error('Registro no encontrado', 404);
         }
 
-        // if (!(new AccessActionPlanActionPolicy())->access($actionPlanAction)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para modificar este registro',
-        //         403
-        //     );
-        // }
+        if (!(new AccessActionPlanPolicy())->access($actionPlanAction->actionPlan)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para modificar este registro',
+                403
+            );
+        }
 
         $response = $this->service->update($request->validated(), $id);
 
@@ -116,12 +151,12 @@ class ActionPlanActionController extends Controller
             return ResponseFormatter::error('Registro no encontrado', 404);
         }
 
-        // if (!(new AccessActionPlanActionPolicy())->access($actionPlanAction)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para modificar este registro',
-        //         403
-        //     );
-        // }
+        if (!(new AccessActionPolicy())->access($actionPlanAction->actionPlan)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para modificar este registro',
+                403
+            );
+        }
 
         $response = $this->service->partialUpdate($request->validated(), $id);
 
@@ -144,12 +179,12 @@ class ActionPlanActionController extends Controller
             return ResponseFormatter::error('Registro no encontrado', 404);
         }
 
-        // if (!(new AccessActionPlanActionPolicy())->access($actionPlanAction)) {
-        //     return ResponseFormatter::error(
-        //         'Usted no tiene autorización para eliminar este registro',
-        //         403
-        //     );
-        // }
+        if (!(new AccessActionPlanPolicy())->access($actionPlanAction->actionPlan)) {
+            return ResponseFormatter::error(
+                'Usted no tiene autorización para eliminar este registro',
+                403
+            );
+        }
 
         $response = $this->service->delete($id);
